+++ lib/models/weather_condition.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:flutter_training/gen/assets.gen.dart';
-import 'package:json_annotation/json_annotation.dart';
 
 enum WeatherCondition {
   sunny,
@@ -39,20 +38,3 @@ extension WeatherConditionExt on WeatherCondition {
   }
 }
 
-class WeatherConditionConverter
-    implements JsonConverter<WeatherCondition?, String?> {
-  const WeatherConditionConverter();
-
-  @override
-  WeatherCondition? fromJson(String? json) {
-    if (json == null) {
-      return null;
-    }
-    return WeatherConditionExt.fromString(json);
-  }
-
-  @override
-  String? toJson(WeatherCondition? object) {
-    return object?.toString().split('.').last;
-  }
-}
